Now, it calls the script "graph.jl".
We have added the object to save the weights
independently of the nodes such that the
reading has a sort of identifier.

Then, we get the weights of the edges considering
the suitable transformation to an integer from
a string.

Again, we need to get the weights when reading the
edges (line 202).

According to the new structure, we change the way
are added and use the initial script in lines
213 - 219.

Then, we return both "edges_brut" and "weights"
objects (line 225).

Within the plot_graph function we now consider the
initial script regarding the new structure in use.